move constants to common module
